+++ CHANGELOG.md
@@ -17,6 +17,9 @@
 - Add startup for `fcgiwrap.service` because sometimes it did not start automatically.
   ([#657](https://github.com/chatmail/relay/pull/657))
 
+- Add `cmdeploy init --force` command for recreating chatmail.ini
+  ([#656](https://github.com/chatmail/relay/pull/656))
+
 - Increase maxproc for reinjecting ports from 10 to 100
   ([#646](https://github.com/chatmail/relay/pull/646))